Example of using a sparse TI

+++ scikit-mps/mpslib/mpslib.py
@@ -371,9 +371,9 @@ class mpslib:
         from multiprocessing import cpu_count
         import time
 
-        #Ncpu = np.int(cpu_count()/1)
-        Ncpu = np.int(np.ceil(cpu_count()*.8))
-        #Ncpu = np.int(cpu_count()/2)
+        #Ncpu = np.int8(cpu_count()/1)
+        Ncpu = np.int8(np.ceil(cpu_count()*.8))
+        #Ncpu = np.int8(cpu_count()/2)
         
         
         # make sure hard data, soft data and mask data are given as variables
@@ -871,7 +871,7 @@ class mpslib:
 
         N = self.xx.size
 
-        N_hard = np.int(np.ceil((1 - step) * N))
+        N_hard = np.int16(np.ceil((1 - step) * N))
 
         i_hard = np.random.choice(N, N_hard)
